+++ modules/Food_Service/TransactionsReport.php
@@ -15,7 +15,7 @@ if ( isset( $_REQUEST['day_start'] )
 if ( empty( $start_date ) )
 {
 	$_REQUEST['day_start'] = '01';
-	$_REQUEST['month_start'] = mb_strtoupper(date('M'));
+	$_REQUEST['month_start'] = date('m');
 	$_REQUEST['year_start'] = date('Y');
 	$start_date = $_REQUEST['day_start'].'-'.$_REQUEST['month_start'].'-'.$_REQUEST['year_start'];
 }
@@ -35,7 +35,7 @@ if ( isset( $_REQUEST['day_end'] )
 if ( empty( $end_date ) )
 {
 	$_REQUEST['day_end'] = date('d');
-	$_REQUEST['month_end'] = mb_strtoupper(date('M'));
+	$_REQUEST['month_end'] = date('m');
 	$_REQUEST['year_end'] = date('Y');
 	$end_date = $_REQUEST['day_end'].'-'.$_REQUEST['month_end'].'-'.$_REQUEST['year_end'];
 }